fix: add system_variable api for rpc
(cherry picked from commit e0dac72001100a96b6b0f0953610e551e878b804)
